refactor : 패키지 구조 변경
- message 클래스의 패키지 변경
